refactoring by moving some exported functions around
