+++ src/components/charts/geo-locations.tsx
@@ -4,8 +4,9 @@ import { Box, Grid, Table } from "@mantine/core"
 import { type QueryResultRow } from "../../types"
 import { DisplayError, DisplayWarning } from "./alerts"
 import { ChartContainer } from "./container"
+import { IntervalOptions } from "./interval-options"
 import { Loading } from "./loading"
-import { IntervalOptions, RowsOptions } from "./options"
+import { RowsOptions } from "./rows-options"
 import { useInterval } from "./use-interval"
 import { useQuery } from "./use-query"
 import { useRows } from "./use-rows"
