+++ src/components/query/took.tsx
@@ -1,0 +1,6 @@
+export function Took({ seconds }: { seconds: number }) {
+  if (seconds < 1) {
+    return <span>{(seconds * 1000).toFixed(1)} milliseconds</span>
+  }
+  return <span>{seconds.toFixed(2)} seconds</span>
+}
