+++ src/components/signed-in.tsx
@@ -1,27 +1,11 @@
 import { Alert, Button, LoadingOverlay } from "@mantine/core"
 import type { ReactNode } from "react"
 
-import { useUserData } from "../contexts/user-context"
+import { useUserData } from "../hooks/use-userdata"
 import { Authenticate } from "./authenticate"
 
 export function SignedIn({ children }: { children: ReactNode }) {
-  // const { push, asPath } = useRouter()
-  // const [location, setLocation] = useLocation()
   const { userData, userError } = useUserData()
-
-  // const isNotSignedIn = userData && !userData.user
-  // useEffect(() => {
-  //   // console.warn("NOT SIGNED IN, NEEDS TO REDIRECT")
-  //   if (isNotSignedIn && !location.startsWith("/signin")) {
-  //     setLocation(`/signin?${new URLSearchParams({ cf: location })}`)
-  //   }
-  // }, [isNotSignedIn, location, setLocation])
-  //
-  // if (!isNotSignedIn && userData !== null) {
-  //   let redirectTo = "/signin"
-  //   redirectTo += `?${new URLSearchParams({ cf: location })}`
-  //   return <Redirect to={redirectTo} />
-  // }
 
   if (userError) {
     return (
