+++ src/components/simple-nav.tsx
@@ -1,7 +1,7 @@
 import { Grid, Text } from "@mantine/core"
 import { Link } from "wouter"
 
-import { useUserData } from "../contexts/user-context"
+import { useUserData } from "../hooks/use-userdata"
 
 export function Nav() {
   const { userData } = useUserData()
