+++ src/hooks/use-userdata.ts
@@ -1,0 +1,7 @@
+import { useContext } from "react"
+
+import { UserDataContext } from "../contexts/user-context"
+
+export function useUserData() {
+  return useContext(UserDataContext)
+}
